+++ src/sourcemaps/tools/sentry-cli.ts
@@ -8,11 +8,11 @@ import {
   abortIfCancelled,
   addSentryCliRc,
   getPackageDotJson,
-  hasPackageInstalled,
   installPackage,
 } from '../../utils/clack-utils';
 
 import { SourceMapUploadToolConfigurationOptions } from './types';
+import { hasPackageInstalled } from '../../utils/package-json';
 
 export async function configureSentryCLI(
   options: SourceMapUploadToolConfigurationOptions,