+++ bin.ts
@@ -4,6 +4,7 @@ import { run } from './lib/Setup';
 import { runNextjsWizard } from './src/nextjs/nextjs-wizard';
 import { runSourcemapsWizard } from './src/sourcemaps/sourcemaps-wizard';
 import { runSvelteKitWizard } from './src/sveltekit/sveltekit-wizard';
+import { runAppleWizard } from './src/apple/apple-wizard';
 import { withTelemetry } from './src/telemetry';
 import { WizardOptions } from './src/utils/types';
 export * from './lib/Setup';
@@ -89,6 +90,16 @@ switch (argv.i) {
       // eslint-disable-next-line no-console
     ).catch(console.error);
     break;
+  case 'ios':
+    withTelemetry(
+      {
+        enabled: !argv['disable-telemetry'],
+        integration: 'ios'
+      },
+      () => runAppleWizard(wizardOptions),
+      // eslint-disable-next-line no-console
+    ).catch(console.error);
+    break
   default:
     void run(argv);
 }