+++ src/utils/clack-utils.ts
@@ -126,10 +126,16 @@ export async function confirmContinueEvenThoughNoGitRepo(): Promise<void> {
   }
 }
 
+export async function askToInstallSentryCLI(): Promise<boolean> {
+  return await abortIfCancelled(clack.confirm({
+    message: "You don't have Sentry CLI installed. Do you want to install it?"
+  }));
+}
+
 export async function askForWizardLogin(options: {
   url: string;
   promoCode?: string;
-  platform?: 'javascript-nextjs' | 'javascript-sveltekit';
+  platform?: 'javascript-nextjs' | 'javascript-sveltekit' | 'apple-ios';
 }): Promise<WizardProjectData> {
   Sentry.setTag('has-promo-code', !!options.promoCode);
 
@@ -182,8 +188,7 @@ export async function askForWizardLogin(options: {
   const urlToOpen = loginUrl.toString();
   clack.log.info(
     `${chalk.bold(
-      `If the browser window didn't open automatically, please open the following link to ${
-        hasSentryAccount ? 'log' : 'sign'
+      `If the browser window didn't open automatically, please open the following link to ${hasSentryAccount ? 'log' : 'sign'
       } into Sentry:`,
     )}\n\n${chalk.cyan(urlToOpen)}`,
   );
@@ -228,6 +233,23 @@ export async function askForWizardLogin(options: {
   Sentry.setTag('opened-wizard-link', true);
 
   return data;
+}
+
+export async function askForItemSelection(items: string[], message: string): Promise<{ value: string, index: number }> {
+  const selection: { value: string, index: number } | symbol = await abortIfCancelled(
+    windowedSelect({
+      maxItems: 12,
+      message: message,
+      options: items.map((item, index) => {
+        return {
+          value: { value: item, index: index },
+          label: item,
+        };
+      }),
+    }),
+  );
+
+  return selection;
 }
 
 export async function askForProjectSelection(
@@ -356,9 +378,8 @@ export async function askForSelfHosted(urlFromArgs?: string): Promise<{
       tmpUrlFromArgs ||
       (await abortIfCancelled(
         clack.text({
-          message: `Please enter the URL of your ${
-            urlFromArgs ? '' : 'self-hosted '
-          }Sentry instance.`,
+          message: `Please enter the URL of your ${urlFromArgs ? '' : 'self-hosted '
+            }Sentry instance.`,
           placeholder: 'https://sentry.io/',
         }),
       ));