+++ src/apple/templates.ts
@@ -16,11 +16,11 @@ export function getSwiftSnippet(dsn: string): string {
             options.debug = true // Enabled debug when first installing is always helpful
             options.enableTracing = true 
 
-            //Uncomment the following lines to add more data to your events
-            //options.attachScreenshot = true //This will add a screenshot to the error events
-            //options.attachViewHierarchy = true //This will add the view hierarchy to the error events
+            // Uncomment the following lines to add more data to your events
+            // options.attachScreenshot = true // This adds a screenshot to the error events
+            // options.attachViewHierarchy = true // This adds the view hierarchy to the error events
         }
-        //Remove the next line after confirming that your Sentry integration is working.
+        // Remove the next line after confirming that your Sentry integration is working.
         SentrySDK.capture(message: "This app uses Sentry! :)")\n`;
 }