+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## Unreleased
 
-- feat: Add support for iOS (#334)
+- feat(apple): Add support for iOS (#334)
+- feat(sourcemaps): Detect SvelteKit and NextJS projects and redirect to dedicated wizards (#341)
+- ref(sourcemaps): Improve Outro message (#344) 
 
 ## 3.5.0