+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - feat(apple): Add support for iOS (#334)
 - feat(apple): iOS wizard has support for cocoapods
+- feat(sourcemaps): Detect SvelteKit and NextJS projects and redirect to dedicated wizards (#341)
+- ref(sourcemaps): Improve Outro message (#344) 
 
 ## 3.5.0