+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - feat(apple): Add support for iOS (#334)
 - feat(apple): iOS wizard has support for cocoapods (#350)
+- feat(sourcemaps): Add CLI-based flow for Angular (#349)
 - feat(sourcemaps): Detect SvelteKit and NextJS projects and redirect to dedicated wizards (#341)
 - ref(sourcemaps): Improve Outro message (#344)