+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - feat(sourcemaps): Add CLI-based flow for Angular (#349)
 - feat(sourcemaps): Detect SvelteKit and NextJS projects and redirect to dedicated wizards (#341)
 - ref(sourcemaps): Improve Outro message (#344) 
+- Use `with-environment.sh` in Xcode Build Phases (#329)
 
 ## 3.5.0