+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - feat(apple): Add support for iOS (#334)
 - feat(sourcemaps): Add CLI-based flow for Angular (#349)
 - feat(sourcemaps): Detect SvelteKit and NextJS projects and redirect to dedicated wizards (#341)
+- feat(sourcemaps): Pre-select auto-detected build tool option (#354)
 - ref(sourcemaps): Improve Outro message (#344) 
 
 ## 3.5.0